fixed warnings

+++ modules/videoio/test/test_audio.cpp
@@ -197,7 +197,7 @@ private:
         {
             if (cap.grab())
             {
-                SCOPED_TRACE(cv::format("frame=%d", videoData.size()+1));
+                SCOPED_TRACE(cv::format("frame=%d", (int)(videoData.size()+1)));
                 ASSERT_TRUE(cap.retrieve(videoFrame));
                 ASSERT_TRUE(cap.retrieve(audioFrame, audioBaseIndex));
                 ASSERT_LT(abs(cap.get(CAP_PROP_AUDIO_POS) -  cap.get(CAP_PROP_POS_MSEC)/ 1000 * samplePerSecond), audioSamplesTolerance);//(double)(videoData.size()+1)
